requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
boto3==1.19.8
botocore==1.22.8
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.8
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==35.0.0
cycler==0.11.0
daphne==3.0.2
Django==3.2.9
django-appconf==1.0.5
django-bootstrap-form==3.4
django-compressor==2.4.1
django-countries==7.2.1
django-cropperjs==0.0.5
django-localflavor==3.1
django-phonenumber-field==6.0.0
django-registration==3.2
django-storages==1.12.3
django-survey-and-report==1.3.35
django-tastypie==0.14.3
djangorestframework==3.12.4
djangorestframework-api-key==2.1.0
et-xmlfile==1.1.0
fonttools==4.28.1
frozenlist==1.2.0
geoip2==4.5.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==0.10.0
kiwisolver==1.3.2
matplotlib==3.5.0rc1
maxminddb==2.2.0
msgpack==1.0.2
multidict==5.2.0
mysqlclient==2.0.3
numpy==1.21.4
opencv-python==4.5.4.60
openpyxl==3.0.9
ordereddict==1.1
packaging==21.2
pandas==1.3.4
phonenumbers==8.12.37
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pySankeyBeta==1.3.0
python-dateutil==2.8.2
python-decouple==3.5
python-mimeparse==1.6.0
python-stdnum==1.17
pytz==2021.3
PyYAML==6.0
rcssmin==1.0.6
requests==2.26.0
rjsmin==1.1.0
s3transfer==0.5.0
scipy==1.7.2
seaborn==0.11.2
service-identity==21.1.0
setuptools-scm==6.3.2
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
twilio==7.4.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
xlwt==1.3.0
yarl==1.7.2
zope.interface==5.4.0
